'i' function macro refactor as function and other minor tweaks

+++ wide/instrument.hpp
@@ -21,7 +21,7 @@ using namespace std;
 
 class Instrument {
 public:
-  Instrument(int id) : id(id),_ch(id) {}
+  Instrument(uint8_t id) : id(id),_ch(id) {}
   
   void play(function<Notes(void)> _f) {
     isWritingPlayFunc->store(true);
@@ -100,7 +100,7 @@ public:
     return _mute;
   }
   
-  int id;
+  uint8_t id;
   uint32_t step = 0, ccStep = 0;
   shared_ptr<function<Notes(void)>> const f = make_shared<function<Notes(void)>>(SILENCE);
   shared_ptr<vector<function<CC(void)>>> const ccs = make_shared<vector<function<CC(void)>>>(NO_CTRL);
